+++ src/DependencyInjection/TaskManagerBundleConfiguration.php
@@ -5,6 +5,9 @@ namespace Torr\TaskManager\DependencyInjection;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
+/**
+ *
+ */
 final class TaskManagerBundleConfiguration implements ConfigurationInterface
 {
 	/**